added index stack to memorize the state of the tabs and added separate pages for each of the tabs
