+++ lib/screens/features/home/community_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class CommunityTab extends StatefulWidget {
+  const CommunityTab({super.key});
+
+  @override
+  State<CommunityTab> createState() => _CommunityTabState();
+}
+
+class _CommunityTabState extends State<CommunityTab> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('Community Page'),
+      ),
+    );
+  }
+}
