+++ lib/screens/features/home/home_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class HomeTab extends StatefulWidget {
+  const HomeTab({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _MyHomeState();
+}
+
+class _MyHomeState extends State<HomeTab> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('Home Page'),
+      ),
+    );
+  }
+}
