+++ lib/screens/features/home/learn_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class LearnTab extends StatefulWidget {
+  const LearnTab({super.key});
+
+  @override
+  State<LearnTab> createState() => _LearnTabState();
+}
+
+class _LearnTabState extends State<LearnTab> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('Learn Page'),
+      ),
+    );
+  }
+}
